requirements.txt
fastapi
uvicorn
postgres
pydantic
sqlalchemy
psycopg2-binary
asyncpg
python-dotenv
alembic
python-jose
passlib[bcrypt]
argon2-cffi
argon2
python-multipart
starlette
pyotp
pyjwt
pydantic[email]
pydantic-settings
loguru
itsdangerous
fastapi-mail
qrcode
pytest
httpx
pytest_asyncio
cryptography